Fix biggest square
We could not use the convex hull algorithm as it would get stuck if
we did not have a convex polygon.

Instead we find the square by taking the shortest distance to the next
point until we have all our points.
We use this to calculate the area of the quad and we then find the
biggest quad from all possible quads.

The quad is then fixed so it can be used in perspective correction.
This is done by taking the most top left corner and then finding the most
bottom left corner and so on and so forth.

This gives us a quad that starts in the top left corner and which points
has an anti-clockwise direction.