UpdateEveryXFramesStage
Now we don't have a specific pipeline, it is instead an interface.

This is used in UpdateEveryXFramesStageBase.
You now add stages directly into the UpdateEveryXFramesStage it then
only calls the update functions on the stages every X frames.

This has been added to fullPipeline to only run fullCornerDetection
every 10 frames.