Add a ThreadedPipelineStage
This stage does NOT support GL stages.
This stage is used in fullCornerDetection to do line detection in a
seperate thread.
This is out of sync so the framerate of line detection is not the same
as the framerate of the rest of the system.
This mitigates the performance penalty of doing line detection.