Add AspectCorrectionStage and LetterboxingStage
AspectCorrectionStage takes an aspect ratio and outputs a framebuffer
that uses that aspect ratio with either width or height being the same
as the input framebuffer.

LetterboxingStage takes an input framebuffer and scales it up to the
screen resolution keeping the aspect ratio of the input framebuffer.
The space around the scaled input is black.