Check the quad from BiggestSquareStage is ok for perspective correction
We now check the biggest quad found if it is perspective correctable.
This means that the angles between the points may not go under 60
degrees and must not exceed 130 degrees.

This is not a real solution to check if the quad is perspective
correction. We are probably over approximating here, however it is more
important that we don't get invalid quads than that we cannot
perspective correct extreme angles.